+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added a function to set resolution for `PostProcessingSprite`.
+
 ## [1.2.1] - 2022-02-11
 ### Fixed
 - Fixed an issue which caused some glTF materials to not load correctly.